refactor: API 응답 상태코드 상세화
