+++ service/catalog/src/main/java/com/bappul/delivery/catalog/domain/repository/StoreSearchViewRepository.java
@@ -6,8 +6,7 @@ import com.bappul.delivery.catalog.application.service.CursorCodec;
 import com.bappul.delivery.catalog.application.service.CursorCodec.Cursor;
 import com.bappul.delivery.catalog.domain.entity.QStoreSearchView;
 import com.bappul.delivery.catalog.domain.entity.Store;
-import com.bappul.delivery.catalog.web.v1.controller.store.StoreSearchCondition;
-import com.bappul.delivery.catalog.web.v1.controller.store.StoreSort;
+import com.bappul.delivery.catalog.web.v1.request.store.StoreSearchCondition;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.BooleanExpression;
