+++ service/catalog/src/main/java/com/bappul/delivery/catalog/web/v1/controller/menu/MenuOwnerController.java
@@ -2,6 +2,7 @@ package com.bappul.delivery.catalog.web.v1.controller.menu;
 
 import com.bappul.delivery.catalog.application.service.MenuService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -10,45 +11,37 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import response.ApiResponse;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/owner/menus")
 public class MenuOwnerController {
 
-  /**
-   * 메뉴 점주 API
-   * [DELETE] 메뉴 삭제 API
-   * [PATCH] 메뉴 품절 토글 API
-   * [PATCH] 메뉴 특정 옵션 품절 토글 API
-   */
-
   private final MenuService menuService;
 
   @DeleteMapping("/{menuId}")
-  public ResponseEntity<ApiResponse<Void>> deleteById(@PathVariable Long menuId) {
+  public ResponseEntity<Void> deleteById(@PathVariable Long menuId) {
     menuService.deleteById(menuId);
-    return ResponseEntity.status(200).body(ApiResponse.success());
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
 
   @PatchMapping("/{menuId}")
-  public ResponseEntity<ApiResponse<Void>> toggleMenuSoldOut(
+  public ResponseEntity<Void> setMenuSoldOut(
       @PathVariable Long menuId,
       @RequestParam("soldOut") boolean soldOut,
       @AuthenticationPrincipal(expression = "claims['uid']") String ownerId)
   {
     menuService.toggleMenuSoldOut(menuId, soldOut, Long.valueOf(ownerId));
-    return ResponseEntity.status(200).body(ApiResponse.success());
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
 
   @PatchMapping("/options/{menuOptionValueId}")
-  public ResponseEntity<ApiResponse<Void>> toggleOptionStatus(
+  public ResponseEntity<Void> setMenuOptionSoldOut(
       @PathVariable Long menuOptionValueId,
       @RequestParam("soldOut") boolean soldOut,
       @AuthenticationPrincipal(expression = "claims['uid']") String ownerId)
   {
     menuService.toggleOptionStatus(menuOptionValueId, soldOut, Long.valueOf(ownerId));
-    return ResponseEntity.status(200).body(ApiResponse.success());
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
 }
