+++ service/catalog/src/main/java/com/bappul/delivery/catalog/web/v1/controller/store/StorePublicController.java
@@ -4,6 +4,7 @@ import com.bappul.delivery.catalog.application.service.MenuService;
 import com.bappul.delivery.catalog.web.v1.response.menu.MenuResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,16 +17,11 @@ import response.ApiResponse;
 @RequestMapping("/api/v1/stores")
 public class StorePublicController {
 
-  /**
-   * 가게 공개 API
-   * [GET] 가게 메뉴 조회
-   */
-
   private final MenuService menuService;
 
   @GetMapping("/{storeId}")
   public ResponseEntity<ApiResponse<List<MenuResponse>>> getMenus(@PathVariable Long storeId) {
     List<MenuResponse> responses = menuService.getMenus(storeId);
-    return ResponseEntity.status(200).body(ApiResponse.success(responses));
+    return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success(responses));
   }
 }
